Move logout_all.go contents into logout.go
